Fix mixer modes from crashing on server
